pyproject.toml
[project]
name = "stones-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "pillow>=11.3.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "django-stubs>=5.2.1",
    "isort>=6.0.1",
    "mypy>=1.16.1",
    "pylint>=3.3.7",
    "pylint-django>=2.6.1",
]
